v1.0 - modified production sample
This is a sample from the working production machine.
Put it all in the home of the admin, create new dir. /home/admin/avamar_quotas/
Edit avamar_tenants.csv - replace with you real tenants names, and assign some quota limits for each tenant.
Edit scripts and replace emails with real ones, so that you can test if emailing works.
Run, or set a crontab as suggested at the end of each script.